+++ api/specs/webserver/openapi-projects-comments.yaml
@@ -195,8 +195,8 @@ components:
       - project_uuid
       - user_id
       - content
-      - created_at
-      - updated_at
+      - created
+      - modified
       type: object
       properties:
         comment_id:
@@ -220,13 +220,13 @@ components:
           title: Content
           type: string
           description: Content of the comment
-        created_at:
-          title: Created At
+        created:
+          title: Created
           type: string
           description: Timestamp on creation
           format: date-time
-        updated_at:
-          title: Updated At
+        modified:
+          title: Modified
           type: string
           description: Timestamp with last update
           format: date-time
@@ -235,31 +235,19 @@ components:
       title: _CreateProjectCommentsBodyParams
       required:
       - content
-      - user_id
       type: object
       properties:
         content:
           title: Content
           type: string
-        user_id:
-          title: User Id
-          exclusiveMinimum: true
-          type: integer
-          minimum: 0
       additionalProperties: false
     _UpdateProjectCommentsBodyParams:
       title: _UpdateProjectCommentsBodyParams
       required:
       - content
-      - user_id
       type: object
       properties:
         content:
           title: Content
           type: string
-        user_id:
-          title: User Id
-          exclusiveMinimum: true
-          type: integer
-          minimum: 0
       additionalProperties: false
